refactor: removed unused code from seemingly complete feature
